requirements.txt
python-dotenv
openai
tqdm